Refacto conftest to isolate repo integration tests
